fix gpg-agent "running already" check
The GPG_ENV file is sourced before doing the gpg-connect-agent check,
but this file (unlike the SSH_ENV file) doesn't export GPG_AGENT_INFO,
so the check always fails.  This results in new gpg-agents continuously
being spawned.

All this commit does is put in the single export to fix the problem.